Migrate RDF-specific tests to shared infrastructure (part 2)
