+++ src/Http/Http.Extensions/test/RequestDelegateGenerator/RequestDelegateCreationTests.Arrays.cs
@@ -480,4 +480,40 @@ app.MapMethods("/hello", new [] { "POST", "PUT" }, (string[] p) => p[0]);
         await VerifyResponseBodyAsync(httpContext, "ValueFromBody");
         await VerifyAgainstBaselineUsingFile(compilation);
     }
+
+    public async Task RequestDelegateHandlesArraysFromExplicitQueryStringSource()
+    {
+        var source = """
+app.MapPost("/", (HttpContext context,
+    [FromHeader(Name = "Custom")] int[] headerValues,
+    [FromQuery(Name = "a")] int[] queryValues,
+    [FromForm(Name = "form")] int[] formValues) =>
+{
+    context.Items["headers"] = headerValues;
+    context.Items["query"] = queryValues;
+    context.Items["form"] = formValues;
+});
+""";
+        var (_, compilation) = await RunGeneratorAsync(source);
+        var endpoint = GetEndpointFromCompilation(compilation);
+        var httpContext = CreateHttpContext();
+
+        httpContext.Request.Query = new QueryCollection(new Dictionary<string, StringValues>
+        {
+            ["a"] = new(new[] { "1", "2", "3" })
+        });
+
+        httpContext.Request.Headers["Custom"] = new(new[] { "4", "5", "6" });
+
+        httpContext.Request.Form = new FormCollection(new Dictionary<string, StringValues>
+        {
+            ["form"] = new(new[] { "7", "8", "9" })
+        });
+
+        await endpoint.RequestDelegate(httpContext);
+
+        Assert.Equal(new[] { 1, 2, 3 }, (int[])httpContext.Items["query"]!);
+        Assert.Equal(new[] { 4, 5, 6 }, (int[])httpContext.Items["headers"]!);
+        Assert.Equal(new[] { 7, 8, 9 }, (int[])httpContext.Items["form"]!);
+    }
 }
